Update types for `SourceUploadCount`

+++ src/commonMain/kotlin/entities/SourceUploadCount.kt
@@ -11,4 +11,4 @@ import kotlin.js.JsExport
  * @property uploadCount The number of uploads originating from the client application.
  */
 @JsExport @Serializable
-data class SourceUploadCount(val sourceName: String? = null, val uploadCount: Int)
+data class SourceUploadCount(val sourceName: String? = null, val uploadCount: Double)
